Manually track x's parent in rb_tree_remove
The algorithm in CLRS does a clever trick which involve storing the
parent of x in the sentinel value's parent pointer when x is a leaf.
This avoids a few edge cases but also makes the code very subtle and
the correctness hard to verify (it's not always clear where x.p gets
set.  Instead, we choose to manually track the parent of x and set a
new x_p variable all three of the places x is set.