Drop the custom assert macro
The original idea was that maybe the asserts were going to be expensive
so you may want to shut some of them off even in a debug build.  It
turns out that there weren't nearly as many asserts as I thought there
were going to be so it's not nearly as much of a concern.